# Speed up serialization and deserialization