--- a/bench/Makefile
+++ b/bench/Makefile
@@ -1,0 +1,2 @@
+default:
+	dune build bench.exe