--- a/bench/float.proto
+++ b/bench/float.proto
@@ -1,0 +1,5 @@
+syntax = "proto3";
+
+message M {
+   float i = 1;
+}