--- a/dune
+++ b/dune
@@ -1,0 +1,4 @@
+(alias
+ (name default)
+ (deps (alias_rec install))
+)