--- a/src/ocaml_protoc_plugin/deserialize.mli
+++ b/src/ocaml_protoc_plugin/deserialize.mli
@@ -1,0 +1,5 @@
+module C = Spec.Deserialize.C
+
+val deserialize: (int * int) list ->
+  ('constr, (int * Field.t) list -> 'a) Spec.Deserialize.compound_list ->
+  'constr -> Reader.t -> 'a