--- a/src/ocaml_protoc_plugin/extensions.mli
+++ b/src/ocaml_protoc_plugin/extensions.mli
@@ -4,5 +4,5 @@ val pp : Format.formatter -> t -> unit
 val show : t -> string
 val equal : t -> t -> bool
 val compare : t -> t -> int
-val get : ('b -> 'b, 'b) Deserialize.S.compound_list -> t -> 'b Result.t
+val get : ('b -> 'b, 'b) Deserialize.S.compound_list -> t -> 'b
 val set : ('a -> Writer.t, Writer.t) Spec.Serialize.compound_list -> t -> 'a -> t