… bench/bench.ml
… bench/bench.proto
… bench/dune
… bench/enum.proto
… bench/enum_list.proto
… bench/float_list.proto
… bench/int64.proto
… bench/int64_list.proto
… bench/plugin/dune
… bench/protoc/dune
… bench/string.proto
… bench/string_list.proto
… src/ocaml_protoc_plugin/field.ml
… src/ocaml_protoc_plugin/infix.ml
… src/ocaml_protoc_plugin/option.ml
… src/ocaml_protoc_plugin/reader.ml
… src/ocaml_protoc_plugin/reader.mli
… src/ocaml_protoc_plugin/result.ml
… src/ocaml_protoc_plugin/result.mli
… src/ocaml_protoc_plugin/serialize.ml
… src/ocaml_protoc_plugin/spec.ml
… src/ocaml_protoc_plugin/writer.ml
… src/ocaml_protoc_plugin/writer.mli
… src/plugin/code.ml
… src/plugin/emit.ml
… src/plugin/parameters.ml
… src/plugin/protoc_gen_ocaml.ml
… src/plugin/scope.ml
… src/plugin/scope.mli
… src/plugin/types.ml