Benchmarking against ocaml-protoc showed a huge number of allocations for both serialization and deserialization, which prompted some work on speeding up serialization and deserialization:
* A quick rewrite to use exceptions instead of the error monad showed a x2 speedup for deserialization
* Serializing directly to the output buffer yields yet another speed improvement for serialization (~x2 speedup)
* Deserialization reads directly from the input buffer to avoid allocations and makes good use of static evaluation. This improves deserialization speed by 2-4x, depending on input. Two strategies for parsing input: 
  * If input is ordered (its recommended to send fields in sorted order), construction is O(1).
  * If input is un-ordered a map is constructed with has O(n log n), and uses ref cells, which hurts the GC (adds roots)
* Improving code generation for extensions and constructors (make), which now uses default values for optional fields rather than using pattern matching to unpack values (potentially reduce GC strain and improve inlining opportunities)
Other changes includes:
  * A bug-fix for extensions, where setting an extension did not remove the previous values in some cases. 
  * Improve emitted code by stripping whitespace at end of line and ensure all files ends with an empty line

For simple structures, ocaml-protoc-plugin is still slower than `ocaml-protoc`, but for more complex structures speed is comparable. 

I'll note that `ocaml-protoc` has chosen to be non-compliant, and uses c stubs for faster encoding and decoding. Some observations on  `ocaml-protoc`:
  * Does not allow repeated enums
  * Sends fields in reverse order (which is allowed but not recommended)
  * Always sends fields, even when fields have default values 
  * Only support packed repeated fields 
  * No proto2 support
  * Does not support extensions