@andersfugmann 2024-01-07
Some benchmarks 

|  Name |  Old - ns/run |  New - ns/run| ratio |
|  -- |  -- |  -- | -- |
| string_list.M/Decode/Plugin | 311040.6960 | 131970.8184 | .42 |
| string_list.M/Decode/Protoc | 69894.1975 | 82798.6758 | 1.18 |
| string_list.M/Encode/Plugin | 233021.2744 | 116567.0649 | .50 |
| string_list.M/Encode/Protoc | 62315.6831 | 76078.9296 | 1.22 |
|  |
|  |
| float_list.M/Decode/Plugin | 76740.8830 | 34555.1781 | .45 |
| float_list.M/Decode/Protoc | 18662.5440 | 18538.2621 | .99 |
| float_list.M/Encode/Plugin | 42953.9669 | 44641.5460 | 1.03 |
| float_list.M/Encode/Protoc | 19822.1616 | 22271.0384 | 1.12 |
|  |
|  |
| int64_list.M/Decode/Plugin | 192442.6824 | 46931.4920 | .24 |
| int64_list.M/Decode/Protoc | 24973.4493 | 25913.0417 | 1.03 |
| int64_list.M/Encode/Plugin | 101449.5188 | 46641.8079 | .45 |
| int64_list.M/Encode/Protoc | 19726.9139 | 20643.0586 | 1.04 |
|  |
|  |
| enum.M/Decode/Plugin | 552.0980 | 274.8362 | .49 |
| enum.M/Decode/Protoc | 40.8793 | 40.1420 | .98 |
| enum.M/Encode/Plugin | 225.4306 | 174.6894 | .77 |
| enum.M/Encode/Protoc | 77.2908 | 71.7628 | .92 |
|  |
|  |
| string.M/Decode/Plugin | 645.4916 | 292.5425 | .45 |
| string.M/Decode/Protoc | 63.0482 | 70.7741 | 1.12 |
| string.M/Encode/Plugin | 286.8491 | 229.2601 | .79 |
| string.M/Encode/Protoc | 88.9577 | 87.3955 | .98 |
|  |
|  |
| float.M/Decode/Plugin | 491.4052 | 247.9175 | .50 |
| float.M/Decode/Protoc | 36.3592 | 40.4689 | 1.11 |
| float.M/Encode/Plugin | 174.0742 | 194.4858 | 1.11 |
| float.M/Encode/Protoc | 71.0880 | 70.6722 | .99 |
|  |
|  |
| int64.M/Decode/Plugin | 579.7528 | 259.3254 | .44 |
| int64.M/Decode/Protoc | 38.2384 | 40.8936 | 1.06 |
| int64.M/Encode/Plugin | 225.6232 | 190.2801 | .84 |
| int64.M/Encode/Protoc | 77.7386 | 71.9856 | .92 |
|  |
|  |
| bench.M/Decode/Plugin | 358543.7805 | 112197.4954 | .31 |
| bench.M/Decode/Protoc | 56696.3379 | 61007.7365 | 1.07 |
| bench.M/Encode/Plugin | 206013.8916 | 112104.4718 | .54 |
| bench.M/Encode/Protoc | 50472.0140 | 52438.8131 | 1.03 |