@andersfugmann 2024-01-11
@AndreasDahl Any suggestions on how we can get this PR merged? I understand that its huge, but I believe the test coverage quite high. 
A suggestion would be to focus on coding style and comment if the libraries needed for running the benchmark (`make bench`) should be added to the opam file.