@andersfugmann 2024-01-21
With the latest optimizations, ocaml-protoc-plugin is now on par (and in some cases actually faster) with protoc:

Numbers below shows relation to `protoc`. A value above one means that Ocaml-protoc-plugin is slower and below means faster. 
The benchmark is run with flambda settings: `-O3 -unbox-closures -remove-unused-arguments -rounds 3 -inline 100.00 -inline-max-depth 3 -inline-max-unroll 3`

| test | protoc | ocaml-protoc-plugin | ratio |
| -- | -- | -- | -- |
| string_list.M/Decode | 69097.7864 | 78123.2441 | 1.1306 |
| string_list.M/Encode | 73026.9203 | 57308.9066 | .7847 |
| float_list.M/Decode | 18099.2257 | 19993.6007 | 1.1046 |
| float_list.M/Encode | 20498.1777 | 11481.1299 | .5601 |
| int64_list.M/Decode | 21890.8594 | 17562.3905 | .8022 |
| int64_list.M/Encode | 18108.9827 | 12021.9094 | .6638 |
| enum.M/Decode | 58.6929 | 53.6873 | .9147 |
| enum.M/Encode | 103.6401 | 98.2260 | .9477 |
| string.M/Decode | 96.4106 | 91.4180 | .9482 |
| string.M/Encode | 115.0315 | 117.5707 | 1.0220 |
| float.M/Decode | 51.4995 | 51.6307 | 1.0025 |
| float.M/Encode | 113.2276 | 94.4504 | .8341 |
| int64.M/Decode | 63.4648 | 51.3960 | .8098 |
| int64.M/Encode | 92.6878 | 95.6652 | 1.0321 |
| bench.M/Decode | 51351.6396 | 49580.7279 | .9655 |
| bench.M/Encode | 52722.1153 | 33273.7429 | .6311 |