@andersfugmann 2024-01-30
@issuu/platypus @AndreasDahl do you have suggestions on how to proceed on this? 

I propose that we either.
1. We agree that as a general rule all code should be reviewed, and you will prioritizing reviewing.
2. We don't really care about having code on master reviewed. Remove the CODEOWERS and I will just push to master at will
3. Transfer the repository to me and you can create / maintain a local fork.

In case I don't get a timely response, I will start merging to main and prepare a new release, but its really not an ideal solution. I would much rather prefer 3. if Issuu has no interest in maintaining these repositories. I would propose the same for ocaml-zmq repository which also have a stale PR. ocaml-zmq repository should be transferred to the ocaml community though.