@andersfugmann 2024-01-30
Note that I really do not prefer solution 2 and think we should avoid if possible.